New: Extract codeframe formatter
